refactor: extract SkillOpt to RobotSail/skillopt, clean up this repo (#1342)
Phase 1 (separate repo): Created RobotSail/skillopt with full commit
history via git filter-repo. Set up pyproject.toml (hatchling + hatch-vcs),
CI (uv-based, pytest + ruff, Python 3.11-3.13 matrix), and renamed
factory.skillopt imports to skillopt.

Phase 2 (this repo cleanup):
- Delete factory/skillopt/ files: reflect, trainer, gate, clip, aggregate,
  failure_tracker, slow_update, skill, types, adapter, adapters/, prompts/,
  __main__
- Keep factory/skillopt/yaml_surface.py and __init__.py (interop seam)
- Split tests/test_skillopt.py: extract yaml_surface tests to
  tests/test_yaml_surface.py, delete original and adapter/integration tests
- Remove skill-refine workflow from definitions.py registry
- Remove skill-refine from WORKFLOW_META in skill_export.py
- Remove SKILL_REVIEWER from AgentRole enum and AGENT_CONFIGS
- Delete factory/agents/prompts/skill_reviewer.md
- Update test assertions in test_lazy_loading.py and
  test_workflow_definitions.py

Closes #1342
