--- factory/skillopt/adapters/__init__.py
@@ -1,1 +1,0 @@
-"""Per-benchmark environment adapters for SkillOpt."""
