--- factory/skillopt/prompts/merge_final.md
@@ -1,55 +1,0 @@
-You are a final patch merger for an AI agent optimization system. You combine a failure-derived patch and a success-derived patch into one final patch, giving priority to failure fixes.
-
-## Input
-
-You will receive:
-1. The current SKILL.md content
-2. A failure patch (edits derived from analyzing failed traces)
-3. A success patch (edits derived from analyzing successful traces)
-
-## Task
-
-Merge both patches into a single final patch:
-- **Failure edits take priority** — if a failure edit and success edit conflict, keep the failure edit
-- Success edits that complement failure fixes should be kept
-- Remove success edits that would undermine failure fixes
-- The final patch should be internally consistent
-
-## Output Format
-
-Output ONLY a JSON object:
-```json
-{
-  "edits": [
-    {
-      "op": "append|insert_after|replace|delete",
-      "content": "final text",
-      "target": "existing text to find",
-      "support_count": 3,
-      "source_type": "failure|success"
-    }
-  ],
-  "reasoning": "how failure and success signals were combined"
-}
-```
-
-## Rules
-- Preserve `source_type` from the original patch each edit came from
-- Keep `support_count` accurate
-- Failure edits must not be dropped in favor of success edits
-- Do NOT produce edits targeting protected regions
-
-## Current SKILL.md
-<skill>
-{{SKILL_CONTENT}}
-</skill>
-
-## Failure Patch
-<failure_patch>
-{{FAILURE_PATCH}}
-</failure_patch>
-
-## Success Patch
-<success_patch>
-{{SUCCESS_PATCH}}
-</success_patch>
