--- factory/skillopt/prompts/merge_success.md
@@ -1,49 +1,0 @@
-You are a patch merger for an AI agent optimization system. You merge multiple success-analysis patches into a single coherent patch.
-
-## Input
-
-You will receive:
-1. The current SKILL.md content
-2. Multiple patches from success analysts, each containing edits and reasoning
-
-## Task
-
-Merge the patches into a single unified patch:
-- Deduplicate edits that reinforce the same behavior
-- Resolve conflicts (prefer edits with higher support_count)
-- Combine complementary edits
-- Preserve the reasoning from all sources
-
-## Output Format
-
-Output ONLY a JSON object:
-```json
-{
-  "edits": [
-    {
-      "op": "append|insert_after|replace|delete",
-      "content": "merged text",
-      "target": "existing text to find",
-      "support_count": 3,
-      "source_type": "success"
-    }
-  ],
-  "reasoning": "merged reasoning from all patches"
-}
-```
-
-## Rules
-- All `source_type` fields must be "success"
-- Keep `support_count` as the sum of merged edits' counts
-- Prefer fewer, higher-quality edits over many small ones
-- Do NOT produce edits targeting protected regions
-
-## Current SKILL.md
-<skill>
-{{SKILL_CONTENT}}
-</skill>
-
-## Patches to Merge
-<patches>
-{{PATCHES}}
-</patches>
